+++ TestProjects/NoHtml/index.ts
@@ -1,0 +1,4 @@
+import * as lib from "./lib";
+
+console.log("Hello");
+lib.run();